requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
atlastk==0.13
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.77.1
future==0.18.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
iso8601==1.0.2
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
optionaldict==0.1.2
orjson==3.6.8
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pypika-tortoise==0.1.4
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a22
sqlmodel==0.0.6
sqlparse==0.4.2
starlette==0.19.1
tortoise==0.1.1
tortoise-orm==0.19.0
typing_extensions==4.2.0
ujson==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
xmltodict==0.13.0
zipp==3.8.0